# [SPARK-49781][SQL] Optimize size of `MapFromArrays` / `MapFromEntries` to be the size of its children