### What changes were proposed in this pull request?
The pr aims to optimize size of `MapFromArrays` / `MapFromEntries` to be the size of its children.


### Why are the changes needed?
In the optimized scenario, the following logic can be hit `ConstantFolding` without the need for exploit data, and the results can be obtained directly.

Inspiration comes from: https://github.com/apache/spark/pull/30504

### Does this PR introduce _any_ user-facing change?
No.


### How was this patch tested?
- Pass GA.
- Manually check.


### Was this patch authored or co-authored using generative AI tooling?
No.